# Optimize the `flutter-tizen run` execution time